backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
redis==5.2.0
pandas==2.2.3
numpy==1.26.4
scipy==1.14.1
statsmodels==0.14.4
aiohttp==3.11.0
python-dotenv==1.0.1
pyarrow==18.0.0
ccxt==4.4.0
websockets==14.0
pydantic==2.10.0
